requirements.txt
django
psycopg2-binary
gunicorn
django-tailwind
whitenoise
python-dotenv